components/pygame-backend/requirements.txt
Flask==2.0.2
Flask-CORS==3.0.10
Pillow==8.3.2
pygame==2.4.0